+++ Sources/Scout/Models/PathExplorer/PathExplorer+ConvenienceFunctions.swift
@@ -5,9 +5,17 @@
 
 import Foundation
 
+// MARK: - Properties
+
 public extension PathExplorer {
 
-    // MARK: - Get
+    /// `real` property for convenience naming
+    var double: Double? { real }
+}
+
+// MARK: - Get
+
+public extension PathExplorer {
 
     /// Get the key at the given path, specified as variadic `PathElement`s
     ///
@@ -29,8 +37,11 @@ public extension PathExplorer {
     func get<T: KeyAllowedType>(_ elements: PathElement..., as type: KeyType<T>) throws -> T {
         try get(Path(elements), as: type)
     }
+}
 
-    // MARK: - Set
+// MARK: - Set
+
+public extension PathExplorer {
 
     /// Set the value of the key at the given path, specified as variadic `PathElement`s
     ///
@@ -68,8 +79,11 @@ public extension PathExplorer {
     mutating func set(_ elements: PathElement..., keyNameTo newKeyName: String) throws {
         try set(Path(elements), keyNameTo: newKeyName)
     }
+}
 
-    // MARK: - Delete
+// MARK: - Delete
+
+public extension PathExplorer {
 
     /// Delete the key at the given path.
     ///
@@ -92,8 +106,11 @@ public extension PathExplorer {
     mutating func delete(_ elements: PathElement..., deleteIfEmpty: Bool = false) throws {
         try delete(Path(elements), deleteIfEmpty: deleteIfEmpty)
     }
+}
 
-    // MARK: - Add
+// MARK: - Add
+
+public extension PathExplorer {
 
     /// Add a value at the given path, specified as variadic `PathElement`s
     ///
@@ -125,8 +142,11 @@ public extension PathExplorer {
     mutating func add<Type: KeyAllowedType>(_ newValue: Any, at elements: PathElement..., as type: KeyType<Type>) throws {
         try add(newValue, at: Path(elements), as: type)
     }
+}
 
-    // MARK: - Paths
+// MARK: - Paths list
+
+extension PathExplorer {
 
     /// Returns all the paths leading to single or group values
     /// - Parameters:
@@ -142,5 +162,27 @@ public extension PathExplorer {
     ///   - filter: Optionnally provide a filter on the key and/or value. Default is `noFilter`
     func listPaths(startingAt elements: PathElement..., filter: PathsFilter = .noFilter) throws -> [Path] {
         try listPaths(startingAt: Path(elements), filter: .noFilter)
+    }
+}
+
+// MARK: - Export
+
+extension PathExplorer {
+
+    var defaultCSVSeparator: String { ";" }
+
+    /// Export the path explorer value to a CSV if possible. Use the default separator ';' if none specified
+    public func exportCSV() throws -> String {
+        try exportCSV(separator: nil)
+    }
+
+    /// Export the path explorer value to the specified format data
+    public func exportData(to format: DataFormat) throws -> Data {
+        try exportData(to: format, rootName: nil)
+    }
+
+    /// Export the path explorer value to the specified format string data
+    public func exportString(to format: DataFormat) throws -> String {
+        try exportString(to: format, rootName: nil)
     }
 }
