+++ Sources/Scout/Models/PathExplorer/PathExplorer+Literal.swift
@@ -3,9 +3,8 @@
 // Copyright (c) Alexis Bridoux 2020
 // MIT license, see LICENSE file for details
 
-import Foundation
-
 extension PathExplorer {
+
     public init(stringLiteral value: Self.StringLiteralType) {
         self.init(value: value)
     }
